refactor game.h header #1
