--- src/game/helper.h
@@ -1,7 +1,0 @@
-#ifndef STASE_HELPER_H
-#define STASE_HELPER_H
-
-void display_control_counts(const Board &);
-void print_feature_frames(const FeatureFrame *);
-
-#endif //STASE_HELPER_H
