+++ src/search/search.cpp
@@ -2,6 +2,8 @@
 #include "search.h"
 #include "thread.h"
 #include "search_tools.h"
+#include "../game/gamestate.hpp"
+#include "../game/eval.hpp"
 
 #include <vector>
 using std::vector;
