+++ src/test/game/cands/test_development_hook.cpp
@@ -1,6 +1,7 @@
 #include "board.h"
 #include "game.h"
 #include "../../test.h"
+#include "../../../game/gamestate.hpp"
 
 #include <vector>
