+++ src/test/game/cands/test_queen_fork_hook.cpp
@@ -4,6 +4,7 @@
 
 #include "../../../game/cands/cands.h"
 #include "../../test.h"
+#include "../../../game/gamestate.hpp"
 
 TestSet<StringTestCase> queen_fork_hook_test_cases = {
         "game-cands-queen-fork-hook",
